+++ src/backend/core/tests/test_models_invitations.py
@@ -79,7 +79,7 @@ def test_models_invitations_is_expired():
         assert expired_invitation.is_expired is True
 
 
-def test_models_invitationd_new_userd_convert_invitations_to_accesses():
+def test_models_invitations_new_user_convert_invitations_to_accesses():
     """
     Upon creating a new user, invitations linked to the email
     should be converted to accesses and then deleted.